Route all make_pr_* publishers through one shared branch-publication pipeline so there is a single PR-publication code path.
