+++ forge/git_scripts/make_pr_java_run_fix.py
@@ -4,8 +4,7 @@
 # work. If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.
 
 import argparse
-import os
-import subprocess
+import shutil
 import sys
 
 from git_scripts.common_git import (
@@ -13,66 +12,29 @@ from git_scripts.common_git import (
     gh,
     parse_coordinate_parts,
     get_origin_owner,
-    stage_and_commit as stage_and_commit_common,
     find_issue_for_coordinates as find_issue_common,
     get_model_display_name,
     get_agent_name,
     format_stats_diff,
     format_forge_revision_section,
     assert_no_dynamic_access_category_regressions,
-    build_ai_branch_name,
-    delete_remote_branch_if_exists,
-    run_git_transport,
 )
-from git_scripts.make_pr_javac_fix import generate_diff_text
-from utility_scripts.library_stats import stats_artifact_dir
+from git_scripts.pr_publication import (
+    BASE_BRANCH,
+    REPO,
+    REVIEWERS,
+    generate_diff_text,
+    publish_branch,
+    stage_library_version_paths,
+)
 from utility_scripts.local_ci_verification import (
     HUMAN_INTERVENTION_LABEL,
     LOCAL_CI_VERIFICATION_KEY,
-    fetch_pr_base_ref,
     format_local_ci_verification_pr_section,
     local_ci_requires_human_intervention,
-    run_local_ci_verification,
 )
 from utility_scripts.metrics_writer import read_pending_metrics
 from utility_scripts.repo_path_resolver import resolve_repo_roots
-import shutil
-
-REPO = 'oracle/graalvm-reachability-metadata'
-BASE_BRANCH = 'master'
-REVIEWERS = ["vjovanov", "jormundur00", "kimeta"]
-def resolve_repo_paths(
-        explicit_repo_path: str | None,
-        explicit_metrics_repo_path: str | None,
-):
-    """Resolve repo and metrics paths using provided values or local defaults."""
-    return resolve_repo_roots(
-        explicit_repo_path,
-        explicit_metrics_repo_path,
-    )
-
-
-def stage_and_commit(
-        group: str,
-        artifact: str,
-        library_version: str,
-        coordinates: str,
-        repo_path: str,
-        metrics_repo_path: str | None = None,
-        include_in_repo_metrics: bool = False,
-):
-    """Stage the expected files/directories and commit with the required message."""
-    candidate_paths = [
-        str(os.path.join("tests", "src", group, artifact, library_version)),
-        str(os.path.join("metadata", group, artifact, "index.json")),
-        str(os.path.join("metadata", group, artifact, library_version)),
-        str(os.path.relpath(stats_artifact_dir(repo_path, group, artifact), repo_path)),
-    ]
-    del metrics_repo_path, include_in_repo_metrics
-    candidate_paths = [path for path in candidate_paths if os.path.exists(os.path.join(repo_path, path))]
-
-    commit_message = f"Fixed test for {coordinates}"
-    stage_and_commit_common(candidate_paths, commit_message, cwd=repo_path)
 
 
 def create_pull_request(
@@ -247,9 +209,9 @@ def parse_flags(argv_list):
     """Parse CLI flags and resolve repository paths."""
     parser = build_parser()
     flags = parser.parse_args(argv_list)
-    repo_path, metrics_repo_path = resolve_repo_paths(
-        explicit_repo_path=flags.reachability_metadata_path,
-        explicit_metrics_repo_path=flags.metrics_repo_path,
+    repo_path, metrics_repo_path = resolve_repo_roots(
+        flags.reachability_metadata_path,
+        flags.metrics_repo_path,
     )
     return flags.coordinates, flags.new_version, repo_path, metrics_repo_path, flags.issue_number
 
@@ -259,42 +221,23 @@ def push_current_branch_to_origin(
         new_version: str,
         repo_path: str,
         metrics_repo_path: str | None = None,
-        include_in_repo_metrics: bool = False,
 ):
     """Create a feature branch, stage and commit changes, and push to the remote."""
     group, artifact, old_version = parse_coordinate_parts(old_coordinates)
     new_coordinates = f"{group}:{artifact}:{new_version}"
 
-    branch = build_ai_branch_name(
-        f"fix-java-run-{group}-{artifact}-{new_version}",
-        cwd=repo_path,
-    )
-    delete_remote_branch_if_exists(branch, cwd=repo_path)
-    subprocess.run(
-        ["git", "switch", "-C", branch],
-        check=True,
-        cwd=repo_path,
-    )
-    stage_and_commit(
-        group,
-        artifact,
-        new_version,
-        new_coordinates,
-        repo_path,
-        metrics_repo_path=metrics_repo_path,
-        include_in_repo_metrics=include_in_repo_metrics,
-    )
-    base_ref = fetch_pr_base_ref(repo_path, REPO, BASE_BRANCH)
-    subprocess.run(["git", "rebase", base_ref], check=True, cwd=repo_path)
-    run_local_ci_verification(
+    branch, _ = publish_branch(
         repo_path=repo_path,
+        branch_suffix=f"fix-java-run-{group}-{artifact}-{new_version}",
         coordinates=new_coordinates,
-        base_commit=base_ref,
+        stage=lambda: stage_library_version_paths(
+            group, artifact, new_version, repo_path, f"Fixed test for {new_coordinates}",
+        ),
         metrics_repo_path=metrics_repo_path,
+        after_verification=lambda: assert_no_dynamic_access_category_regressions(
+            repo_path, old_coordinates, new_coordinates,
+        ),
     )
-    assert_no_dynamic_access_category_regressions(repo_path, old_coordinates, new_coordinates)
-
-    run_git_transport(["push", "origin", branch], cwd=repo_path)
 
     return branch, group, artifact, old_version, new_coordinates
