+++ forge/git_scripts/make_pr_javac_fix.py
@@ -4,136 +4,37 @@
 # work. If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.
 
 import argparse
-import os
 import shutil
-import subprocess
 import sys
-import tempfile
 
 from git_scripts.common_git import (
     ensure_gh_authenticated,
     gh,
     parse_coordinate_parts,
     get_origin_owner,
-    git_files_under,
-    is_java_fix_test_module_file,
-    stage_and_commit as stage_and_commit_common,
     find_issue_for_coordinates as find_issue_common,
     get_model_display_name,
     get_agent_name,
     format_stats_diff,
     format_forge_revision_section,
     assert_no_dynamic_access_category_regressions,
-    build_ai_branch_name,
-    delete_remote_branch_if_exists,
-    get_configured_reviewers,
-    run_git_transport,
 )
-from utility_scripts.library_stats import stats_artifact_dir
+from git_scripts.pr_publication import (
+    BASE_BRANCH,
+    REPO,
+    REVIEWERS,
+    generate_diff_text,
+    publish_branch,
+    stage_library_version_paths,
+)
 from utility_scripts.local_ci_verification import (
     HUMAN_INTERVENTION_LABEL,
     LOCAL_CI_VERIFICATION_KEY,
-    fetch_pr_base_ref,
     format_local_ci_verification_pr_section,
     local_ci_requires_human_intervention,
-    run_local_ci_verification,
 )
 from utility_scripts.metrics_writer import read_pending_metrics
 from utility_scripts.repo_path_resolver import resolve_repo_roots
-
-REPO = "oracle/graalvm-reachability-metadata"
-BASE_BRANCH = 'master'
-REVIEWERS = get_configured_reviewers()
-def resolve_repo_paths(
-        explicit_repo_path: str | None,
-        explicit_metrics_repo_path: str | None,
-):
-    """Resolve repo and metrics paths using provided values or local defaults."""
-    return resolve_repo_roots(
-        explicit_repo_path,
-        explicit_metrics_repo_path,
-    )
-
-
-def stage_and_commit(
-        group: str,
-        artifact: str,
-        library_version: str,
-        coordinates: str,
-        repo_path: str,
-        metrics_repo_path: str | None = None,
-        include_in_repo_metrics: bool = False,
-):
-    """Stage the expected files/directories and commit with the required message."""
-    candidate_paths = [
-        str(os.path.join("tests", "src", group, artifact, library_version)),
-        str(os.path.join("metadata", group, artifact, "index.json")),
-        str(os.path.join("metadata", group, artifact, library_version)),
-        str(os.path.relpath(stats_artifact_dir(repo_path, group, artifact), repo_path)),
-    ]
-    del metrics_repo_path, include_in_repo_metrics
-    candidate_paths = [path for path in candidate_paths if os.path.exists(os.path.join(repo_path, path))]
-
-    commit_message = f"Fixed test for {coordinates}"
-    stage_and_commit_common(candidate_paths, commit_message, cwd=repo_path)
-
-
-def _copy_git_files_under(repo_path: str, source_dir: str, destination_dir: str) -> None:
-    os.makedirs(destination_dir, exist_ok=True)
-    for git_path in git_files_under(repo_path, source_dir):
-        source_file = os.path.join(repo_path, git_path)
-        relative_file = os.path.relpath(source_file, source_dir)
-        if relative_file.startswith(os.pardir + os.sep) or relative_file == os.pardir:
-            continue
-        if not is_java_fix_test_module_file(relative_file):
-            continue
-        destination_file = os.path.join(destination_dir, relative_file)
-        os.makedirs(os.path.dirname(destination_file), exist_ok=True)
-        shutil.copy2(source_file, destination_file)
-
-
-def generate_diff_text(group: str, artifact: str, current_version: str, new_version: str, repo_path: str):
-    """Build a diff text between the current and new test directories for a library version."""
-    old_dir = os.path.join(repo_path, "tests", "src", group, artifact, current_version)
-    new_dir = os.path.join(repo_path, "tests", "src", group, artifact, new_version)
-    old_display_dir = os.path.relpath(old_dir, repo_path).replace(os.sep, "/")
-    new_display_dir = os.path.relpath(new_dir, repo_path).replace(os.sep, "/")
-    old_diff_display_dir = f"/{old_display_dir}"
-    new_diff_display_dir = f"/{new_display_dir}"
-
-    # Prepare temp copies from Git's file list with a narrow allow-list for the PR body diff.
-    with tempfile.TemporaryDirectory() as tmpdir:
-        tmp_old = os.path.join(tmpdir, "old")
-        tmp_new = os.path.join(tmpdir, "new")
-
-        _copy_git_files_under(repo_path, old_dir, tmp_old)
-        _copy_git_files_under(repo_path, new_dir, tmp_new)
-
-        # Compute diff text using: git diff --no-index --find-renames tmp_old tmp_new
-        res = subprocess.run(
-            ["git", "diff", "--no-index", "--find-renames", tmp_old, tmp_new],
-            capture_output=True,
-            text=True,
-        )
-        if res.returncode in (0, 1):
-            diff_text = res.stdout
-
-            diff_text = diff_text.replace(tmp_old.replace(os.sep, "/"), old_diff_display_dir)
-            diff_text = diff_text.replace(tmp_new.replace(os.sep, "/"), new_diff_display_dir)
-            diff_text = diff_text.replace(tmp_old, old_diff_display_dir)
-            diff_text = diff_text.replace(tmp_new, new_diff_display_dir)
-            return diff_text
-        else:
-            error_message = (
-                f"Subprocess failed with return code: {res.returncode}. "
-                f"Error output:\n{res.stderr.strip()}"
-            )
-            raise RuntimeError(error_message)
-
-
-def load_fix_javac_metrics(metrics_repo_root: str, old_coordinates: str, new_coordinates: str):
-    """Load the pending metrics entry for a javac-fix run."""
-    return read_pending_metrics(metrics_repo_root)
 
 
 def create_pull_request(
@@ -220,11 +121,7 @@ def build_pull_request_preview(
 ) -> tuple[str, str, dict]:
     """Build the PR title/body without creating a GitHub pull request."""
     issue_no = issue_number if issue_number is not None else find_issue_common(new_coordinates, REPO)
-    metrics_entry = load_fix_javac_metrics(
-        metrics_repo_root=metrics_repo_root,
-        old_coordinates=old_coordinates,
-        new_coordinates=new_coordinates,
-    )
+    metrics_entry = read_pending_metrics(metrics_repo_root)
     metrics = metrics_entry.get("metrics", {})
     strategy_name = metrics_entry.get("strategy_name", "")
     model_display_name = get_model_display_name(strategy_name)
@@ -334,9 +231,9 @@ def parse_flags(argv_list):
     """Parse CLI flags and resolve repository paths."""
     parser = build_parser()
     flags = parser.parse_args(argv_list)
-    repo_path, metrics_repo_path = resolve_repo_paths(
-        explicit_repo_path=flags.reachability_metadata_path,
-        explicit_metrics_repo_path=flags.metrics_repo_path,
+    repo_path, metrics_repo_path = resolve_repo_roots(
+        flags.reachability_metadata_path,
+        flags.metrics_repo_path,
     )
     return flags.coordinates, flags.new_version, repo_path, metrics_repo_path, flags.issue_number
 
@@ -346,42 +243,23 @@ def push_current_branch_to_origin(
         new_version: str,
         repo_path: str,
         metrics_repo_path: str | None = None,
-        include_in_repo_metrics: bool = False,
 ):
     """Create a feature branch, stage and commit changes, and push to the remote."""
     group, artifact, old_version = parse_coordinate_parts(old_coordinates)
     new_coordinates = f"{group}:{artifact}:{new_version}"
 
-    branch = build_ai_branch_name(
-        f"fix-javac-{group}-{artifact}-{new_version}",
-        cwd=repo_path,
-    )
-    delete_remote_branch_if_exists(branch, cwd=repo_path)
-    subprocess.run(
-        ["git", "switch", "-C", branch],
-        check=True,
-        cwd=repo_path,
-    )
-    stage_and_commit(
-        group,
-        artifact,
-        new_version,
-        new_coordinates,
-        repo_path,
-        metrics_repo_path=metrics_repo_path,
-        include_in_repo_metrics=include_in_repo_metrics,
-    )
-    base_ref = fetch_pr_base_ref(repo_path, REPO, BASE_BRANCH)
-    subprocess.run(["git", "rebase", base_ref], check=True, cwd=repo_path)
-    run_local_ci_verification(
+    branch, _ = publish_branch(
         repo_path=repo_path,
+        branch_suffix=f"fix-javac-{group}-{artifact}-{new_version}",
         coordinates=new_coordinates,
-        base_commit=base_ref,
+        stage=lambda: stage_library_version_paths(
+            group, artifact, new_version, repo_path, f"Fixed test for {new_coordinates}",
+        ),
         metrics_repo_path=metrics_repo_path,
+        after_verification=lambda: assert_no_dynamic_access_category_regressions(
+            repo_path, old_coordinates, new_coordinates,
+        ),
     )
-    assert_no_dynamic_access_category_regressions(repo_path, old_coordinates, new_coordinates)
-
-    run_git_transport(["push", "origin", branch], cwd=repo_path)
 
     return branch, group, artifact, old_version, new_coordinates
